refactor: separate repayment plan types
